Start on web Vue components, start fixing shared lib imports, imported 3rd party libs
Turns out Vue does not resolve out-of-scope imports, meaning the shared libraries, I mentioned this on their Discord.
Turns out we have a bunch of imports from /bot in /lib, we need to fix this or the port is dead in the water, started on this process here.
Started on the audio player in the presenter but did bump into build problems due to the above.
Successfully added and used the Discord.js library, also added OBS-Websockets but not tried it yet.