Switch to decorators for data and option classes, change how references are handled
This is a fairly drastic transition, the enlisting and meta data collection is now entirely done per class and field decorators.
It collects all things concerning a field in one place, for larger classes this is really helpful.
It remains to convert the two lagest classes, evend and action, that comes next.
This also means a different way to read files from the database, that comes next next.