Move node info to decorators and ItemMap, switch handle ID to actual IDs not type, will support multiple handles of same type per node-side, update handle pos
Most of this was written on a London holiday, early mornings when everyone else was asleep or still in bed.
The finishing touches and bugfixes were done in the days after arriving home, on the same laptop as this commit was still in limbo.
Updated dependencies
Added Session value for color mode